main-motor-control/main_foc_endat_system/script/xbox-controller/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "xbox-controller"
version = "0.1.0"
description = "Xbox controller input interface for custom inverter motor control"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "pygame>=2.6.1",
    "pyserial>=3.5",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.2.0",
]

[project.scripts]
xbox-ctrl = "main:main"
xbox-dashboard = "dashboard:main"

[tool.setuptools]
py-modules = ["main", "controller", "commands", "mapping", "uart", "runtime", "runtime_models", "dashboard"]
